Rest endpoints service validation, exception handling,log,comment

+++ Green-Shop-Rest/src/main/java/com/example/greenshoprest/service/impl/RatingsreviewServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.greenshopcommon.dto.ratingsreviewDto.CreateRatingsreviewReque
 import com.example.greenshopcommon.dto.ratingsreviewDto.RatingsreviewDto;
 import com.example.greenshopcommon.dto.ratingsreviewDto.UpdateRatingsreviewRequestDto;
 import com.example.greenshopcommon.entity.Ratingsreview;
+import com.example.greenshopcommon.exception.EntityNotFoundException;
 import com.example.greenshopcommon.mapper.RatingsreviewMapper;
 import com.example.greenshopcommon.repository.ProductRepository;
 import com.example.greenshopcommon.repository.RatingsreviewRepository;
@@ -12,6 +13,7 @@ import com.example.greenshoprest.security.CurrentUser;
 import com.example.greenshoprest.service.ProductService;
 import com.example.greenshoprest.service.RatingsreviewService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -22,6 +24,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class RatingsreviewServiceImpl implements RatingsreviewService {
 
     private final RatingsreviewRepository ratingsreviewRepository;
@@ -29,6 +32,7 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
     private final ProductRepository productRepository;
     private final RatingsreviewMapper ratingsreviewMapper;
 
+    // Fetch all ratings reviews by productId and map them to DTOs
     @Override
     public ResponseEntity<List<RatingsreviewDto>> getAllByProductId(int productId) {
         List<RatingsreviewDto> ratingDtoAll = ratingsreviewRepository
@@ -39,6 +43,7 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
         return ResponseEntity.ok(ratingDtoAll);
     }
 
+    // Fetch all ratings reviews and map them to DTOs
     @Override
     public ResponseEntity<List<RatingsreviewDto>> getAll() {
         List<RatingsreviewDto> ratingDtoAll = ratingsreviewRepository
@@ -48,8 +53,11 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
                 .collect(Collectors.toList());
         return ResponseEntity.ok(ratingDtoAll);
     }
+
+    // Create a new ratings review and rating for a product
     @Override
     public ResponseEntity<?> createReviewAndRating(CreateRatingsreviewRequestDto createRatingsreviewRequestDto, CurrentUser currentUser) {
+        log.info("Creating a new ratings review and rating");
         return productRepository.findById(createRatingsreviewRequestDto.getProductDto().getId())
                 .map(product -> {
                     Ratingsreview ratingsreview = ratingsreviewMapper.map(createRatingsreviewRequestDto);
@@ -62,17 +70,32 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
                 })
                 .orElseGet(() -> ResponseEntity.noContent().build());
     }
+
+    // Fetch ratings reviews by user ID and map them to DTOs
     @Override
     public ResponseEntity<?> getRatingsreviewByUserId(CurrentUser currentUser) {
+        if (currentUser == null) {
+            throw new IllegalArgumentException("CurrentUser or its associated User must not be null.");
+        }
+        log.info("Fetching ratings review for user with ID: {}", currentUser.getUser().getId());
         return Optional.ofNullable(currentUser)
                 .map(CurrentUser::getUser)
                 .flatMap(user -> ratingsreviewRepository.findRatingsreviewByUserId(user.getId())
                         .map(ratingsreviewMapper::mapToDto))
                 .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.noContent().build());
+                .orElseThrow(() -> {
+                    log.info("Ratings review not found for user with ID: {}", currentUser.getUser().getId());
+                    return new EntityNotFoundException("Ratings review not found for user with ID " + currentUser.getUser().getId() + ".");
+                });
     }
+
+    // Update an existing ratings review and rating
     @Override
     public ResponseEntity<?> updateRatingsreview(UpdateRatingsreviewRequestDto updateRatingsreviewRequestDto, CurrentUser currentUser) {
+        if (currentUser == null && updateRatingsreviewRequestDto == null) {
+            throw new IllegalArgumentException("CurrentUser and updateRatingsreviewRequestDto must not be null.");
+        }
+        log.info("Updating ratings review with ID: {}", updateRatingsreviewRequestDto.getId());
         return productRepository.findById(updateRatingsreviewRequestDto.getProductDto().getId())
                 .map(product -> {
                     Ratingsreview ratingsreview = ratingsreviewMapper.updateDto(updateRatingsreviewRequestDto);
@@ -83,20 +106,35 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
                     RatingsreviewDto ratingsreviewDto = ratingsreviewMapper.mapToDto(ratingsreview);
                     return ResponseEntity.ok(ratingsreviewDto);
                 })
-                .orElseGet(() -> ResponseEntity.noContent().build());
+                .orElseThrow(() -> {
+                    log.info("Ratings review with ID {} not found for update", updateRatingsreviewRequestDto.getId());
+                    return new EntityNotFoundException("Ratings review with ID " + updateRatingsreviewRequestDto.getId() + " not found.");
+                });
+
     }
+
+    // Delete a ratings review and rating by ID
     @Override
     public ResponseEntity<?> deleteRatingsreview(int id) {
+        if (id <= 0) {
+            throw new IllegalArgumentException("The id cannot be 0 or less than 0: " + id);
+        }
+        log.info("Deleting ratings review with ID: {}", id);
         return ratingsreviewRepository.findById(id)
                 .map(ratingsreview -> {
                     ratingsreviewRepository.deleteById(id);
                     return ResponseEntity.noContent().build();
                 })
-                .orElse(ResponseEntity.notFound().build());
+                .orElseThrow(() -> {
+                    log.info("Ratings review with ID {} not found for deletion", id);
+                    return new EntityNotFoundException("Ratings review with ID " + id + " not found.");
+                });
     }
 
+    // Fetch all products with their associated ratings
     @Override
     public ResponseEntity<List<ProductDto>> allProductsRating() {
+        log.info("Fetching all products with ratings");
         List<ProductDto> products = productService.findProducts().getBody();
         List<ProductDto> productsWithRating = products.stream()
                 .map(productDto -> {
@@ -112,6 +150,7 @@ public class RatingsreviewServiceImpl implements RatingsreviewService {
         return ResponseEntity.ok(productsWithRating);
     }
 
+    // Helper method to calculate the average rating for a product
     public double calculateProductRating(List<RatingsreviewDto> ratingsreviews) {
         int totalRatings = ratingsreviews.size();
         double sumRatings = 0.0;
